Allow .attr-cache in hackGet
This is used by ob thunk to keep gc paths for nix. We need to support
this as well.

This probably can be made less complex, but for now we only have 4
cases to match. If we get more, it might make sense to make optional /
required matches more clear.